Factor out lambda_service.py, add more docs.
